Fix reporte Detallado Pregrado.
